Address bug 2268, concerning unreachable statements in epetra headers. (These statements were in unsupported functions where an abort preceded a return.)
These warnings showed up when building with the Sun Solaris compiler
with the +w2 flag. This flag exposed a bunch of other warnings which
I also fixed, mostly:
-  unused variables,
-  variables hiding variables in an outer scope,
-  derived-class methods hiding methods with the same name in the
   base class

There are some warnings in the headers Epetra_BasicRowMatrix.h and
Epetra_JadMatrix.h which I didn't touch since I believe Mike is still
actively working in those.

All epetra tests run and pass.

+++ packages/epetra/src/Epetra_OffsetIndex.cpp
@@ -240,7 +240,7 @@ void Epetra_OffsetIndex::GenerateRemoteOffsets_( const Epetra_CrsGraph & SourceG
   int * RecvArray = 0;
   int RecvArraySize = 0;
   Distor.Do( reinterpret_cast<char *>(SourceArray),
-             sizeof(int),
+             (int)sizeof(int),
              Sizes,
              RecvArraySize,
              cRecvArray );
